New Office OpenCL format joining the separate ones for Office 2007, Office 2010 and Office 2013. Also move AES decryption and verify to GPU-side. Closes #2419.
